backend/requirements.txt
# RBM Resource Module - Backend dependencies
# Install: pip install -r requirements.txt (from backend/ or project root)

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Auth & security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Config & validation
python-dotenv>=1.0.0
pydantic>=2.0.0

# PDF export (audit logs)
reportlab>=4.0.0
